Create main.py
Remember to put your own app_eui and app_key.
This piece of code reads values from two different DHT21 sensors. Then forwards this data as well as a blind state (open or closed) and real time clock in the microcontroller to The Things Network, which sends to Thingspeak where graphs are plotted.