Payload Format in The Things Network
In order to receive the correct data that was read from the sensor we have to only transmit positive values. In the microcontroller, we sent two values for each parameter: an 8 bit value cointaining the integer of the value returned from the DHT21 function multiplied by 100 and divided by 256 (2^8); and another 8 bit value that contains the rest of the previous division.